% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartisFunctions.R
\name{annotateSequences}
\alias{annotateSequences}
\title{Perform sequence annotation with partis.
This function does call partis.}
\usage{
annotateSequences(input_filename, output_filename = "partis_output.csv",
  partis_path = Sys.getenv("PARTIS_PATH"), num_procs = 4,
  partition = TRUE, collapse_clones = TRUE, cleanup = TRUE,
  do_full_annotation = TRUE, output_path = "_output", run_partis = TRUE,
  germline_dir = NULL)
}
\arguments{
\item{input_filename}{The input (fasta) file, i.e. --infname argument}

\item{output_filename}{Desired output (csv) filename, i.e. --outfname 
argument}

\item{partis_path}{The full path to the partis executable}

\item{num_procs}{The number of processors to use in parallel}

\item{cleanup}{Flag to delete all interim files created by partis}

\item{output_path}{Desired output directory, which will contain 
\code{output_filename}}
}
\value{
A data.table object containing the output of the partis annotate call
}
\description{
Perform sequence annotation with partis.
This function does call partis.
}
