% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartisFunctions.R
\name{annotateSequencesFromStrings}
\alias{annotateSequencesFromStrings}
\title{Run partis annotate on vector or list of sequences directly (i.e. without
  saving as a fasta file)}
\usage{
annotateSequencesFromStrings(sequences, ...)
}
\arguments{
\item{sequences}{List or vector of mature BCR sequences for annotation}
}
\description{
Run partis annotate on vector or list of sequences directly (i.e. without
  saving as a fasta file)
}
