% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{binContinuousListsAsDiscrete}
\alias{binContinuousListsAsDiscrete}
\title{Discretize two lists of continuous data into mutual, well-defined bins.
This function is}
\usage{
binContinuousListsAsDiscrete(list_a, list_b)
}
\arguments{
\item{list_a}{First list to bin}

\item{list_b}{Second list to bin}
}
\value{
a list of vectors containing the corresponding counts for each bin
}
\description{
Discretize two lists of continuous data into mutual, well-defined bins.
This function is
}
\examples{
list_a <- runif(100)
list_b <- runif(100)
binned <- binContinuousListsAsDiscrete(list_a, list_b) \%>\% 
    melt(value.factor=TRUE) \%>\% 
    mutate(bin_number=rep(seq(1:(nrow(.)/2)), 2))
names(binned)[1:2] <- c("count", "list_id")
ggplot(binned, aes(x=bin_number, y=count, fill=as.factor(list_id))) + 
    geom_bar(stat="identity", position="dodge")
}
