% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare.R
\name{bootstrapFasta}
\alias{bootstrapFasta}
\title{Resample DNA sequences from a fasta file with replacement, and save the
  resultant sequences to a new fasta file}
\usage{
bootstrapFasta(fasta_file, output_filename)
}
\arguments{
\item{fasta_file}{Fasta file from which to resample sequences}

\item{output_filename}{Desired output_filename}
}
\description{
This acts as a bootstrap on a DNA sequence dataset as an attempt to get 
  a benchmark 'control' for assessing similarity in comparison functions.
  See \code{compareRepertoires} for more details.
}
