% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{checkColumn}
\alias{checkColumn}
\title{A wrapper of \code{alakazam::checkColumns} that stops execution when an
  error is thrown.
Checks whether \code{column} is a column name present in \code{dat},
  and whether this column contains at least some non-missing values.}
\usage{
checkColumn(dat, column)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{A string naming the expected column of \code{dat}}
}
\description{
A wrapper of \code{alakazam::checkColumns} that stops execution when an
  error is thrown.
Checks whether \code{column} is a column name present in \code{dat},
  and whether this column contains at least some non-missing values.
}
