% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{collapseAlleles}
\alias{collapseAlleles}
\title{Ignore allelic differences of common genes}
\usage{
collapseAlleles(gene_list)
}
\arguments{
\item{gene_list}{Vector of genes possibly broken down by allele}
}
\value{
Vector of genes with allelic variant information discarded
}
\description{
This function is used when reporting gene usage, so that two alleles of a
common gene are both treated as the one gene, and not separately.
For example, IGHV3-13*01 and IGHV3-13*02 both become IGHV3-13.
}
