% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareAcidityDistributions}
\alias{compareAcidityDistributions}
\title{Compare the acidity distributions of two datasets}
\usage{
compareAcidityDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\description{
Compare the acidity distributions of two datasets
}
