% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareAliphaticIndexDistributions}
\alias{compareAliphaticIndexDistributions}
\title{Compare the distributions of aliphatic indices of two datasets}
\usage{
compareAliphaticIndexDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a}{First dataset, a data.table or data.frame}

\item{dat_b}{Second datset, a data.table or data.frame}
}
\value{
The JS divergence of the two distributions
}
\description{
Compare the distributions of aliphatic indices of two datasets
}
