% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareAromaticityDistributions}
\alias{compareAromaticityDistributions}
\title{Compare the aromaticity distributions of two datasets}
\usage{
compareAromaticityDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\description{
Compare the aromaticity distributions of two datasets
}
