% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareAtchleyFactorDistributions}
\alias{compareAtchleyFactorDistributions}
\title{Compare the distributions of mean Atchley factors for two datasets}
\usage{
compareAtchleyFactorDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a}{First dataset, a data.table or data.frame}

\item{dat_b}{Second dataset, a data.table or data.frame}
}
\value{
JS divergence of Atchley factor distributions
}
\description{
Compare the distributions of mean Atchley factors for two datasets
}
