% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareBasicityDistributions}
\alias{compareBasicityDistributions}
\title{Compare the basicity distributions of two datasets}
\usage{
compareBasicityDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\description{
Compare the basicity distributions of two datasets
}
