% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareBulkinessDistributions}
\alias{compareBulkinessDistributions}
\title{Compare the bulkiness distributions of two datasets}
\usage{
compareBulkinessDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\description{
Compare the bulkiness distributions of two datasets
}
