% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareCategoricalDistributions}
\alias{compareCategoricalDistributions}
\title{Similarity metric for joint categorical distributions}
\usage{
compareCategoricalDistributions(table_a, table_b)
}
\arguments{
\item{table_a}{First contingency table}

\item{table_b}{Second contingency table}
}
\value{
A positive-valued distance metric reflecting the similarity of the
  distributions manifest in table_a and table_b
}
\description{
Currently unused. This metric attempts to compare contingency tables of two
  distributions that accounts for joint and all marginal usage rates.
}
