% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareCDR3Distributions}
\alias{compareCDR3Distributions}
\title{Compare levenshtein distance distributions of two CDR3 repertoires}
\usage{
compareCDR3Distributions(dat_a, dat_b, subsample = TRUE,
  subsample_count = 100, trial_count = 10)
}
\arguments{
\item{dat_a}{First dataset, a data.table or data.frame object}

\item{dat_b}{Second dataset, a data.table or data.frame object}
}
\value{
The JS divergence of the levenshtein distance distributions of the
  CDR3s of the two repertoires
}
\description{
Compare levenshtein distance distributions of two CDR3 repertoires
}
