% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareCDR3Lengths}
\alias{compareCDR3Lengths}
\title{Compare the distribution of CDR3 lengths for two datasets}
\usage{
compareCDR3Lengths(dat_a, dat_b)
}
\arguments{
\item{dat_a}{First dataset}

\item{dat_b}{Second dataset}
}
\value{
The JS divergence of the two CDR3 length distributions
}
\description{
Compare the distribution of CDR3 lengths for two datasets
}
