% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareChargeDistributions}
\alias{compareChargeDistributions}
\title{Compare the charge distributions of two datasets}
\usage{
compareChargeDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\description{
Compare the charge distributions of two datasets
}
