% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareColdspotCounts}
\alias{compareColdspotCounts}
\title{Compare coldspot count distributions of two sets of mature BCR sequences}
\usage{
compareColdspotCounts(dat_a, dat_b, subsample_count = 1000,
  trial_count = 10)
}
\arguments{
\item{dat_a}{First dataset}

\item{dat_b}{Second dataset}

\item{subsample_count}{Number of subsamples for averaging}

\item{trial_count}{Number of subsample trials}
}
\value{
The JS divergence of the coldspot count distributions inferred from
  \code{dat_a$mature_seq} and \code{dat_b$mature_seq}, respectively
}
\description{
Compare coldspot count distributions of two sets of mature BCR sequences
}
