% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareCollessLikeIndices}
\alias{compareCollessLikeIndices}
\title{Compare Colless-like indices for two phylogenetic trees}
\usage{
compareCollessLikeIndices(tree_1, tree_2)
}
\arguments{
\item{tree_1}{The first phylo object}

\item{tree_2}{The second phylo object}
}
\value{
The absolute difference of the respective Colless-like indices
}
\description{
Compare Colless-like indices for two phylogenetic trees
}
