% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareCopheneticIndices}
\alias{compareCopheneticIndices}
\title{Compare cophenetic indices for two phylogenetic trees}
\usage{
compareCopheneticIndices(tree_1, tree_2)
}
\arguments{
\item{tree_1}{The first phylo object}

\item{tree_2}{The second phylo object}
}
\value{
The absolute difference of the respective cophenetic indices
}
\description{
Compare cophenetic indices for two phylogenetic trees
}
