% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareDeletionLengths}
\alias{compareDeletionLengths}
\title{Compare the deletion length distributions of two datasets}
\usage{
compareDeletionLengths(dat_a, dat_b, gene, end)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}

\item{gene}{Gene label Either "VGene", "DGene", or "JGene".}

\item{end}{The end of gene over which the deletion occured. Either "3Prime" 
or "5Prime".}
}
\value{
JS divergence of deletion lengths between \code{dat_a} and \code{dat_b}
}
\description{
Compare the deletion length distributions of two datasets
}
