% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareDGene3PrimeDeletionLengthDistributions}
\alias{compareDGene3PrimeDeletionLengthDistributions}
\title{Compare the D gene 3' deletion length distributions of two datasets}
\usage{
compareDGene3PrimeDeletionLengthDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a}{A \code{data.table} corresponding to repertoire annotations}

\item{dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\description{
Compare the D gene 3' deletion length distributions of two datasets
}
