% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareDistanceFromNaiveToMatureDistributions}
\alias{compareDistanceFromNaiveToMatureDistributions}
\title{Compare Levenshtein distance distributions from naive sequences to}
\usage{
compareDistanceFromNaiveToMatureDistributions(dat_a, dat_b,
  approximate = FALSE, v_gene_only = FALSE)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}

\item{approximate}{If TRUE, approximate distribution by subsampling.}

\item{v_gene_only}{If TRUE, restrict sequences to the V gene only}
}
\description{
Compare Levenshtein distance distributions from naive sequences to
}
