% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareDistancesFromNaiveToMature}
\alias{compareDistancesFromNaiveToMature}
\title{Compare Levenshtein distance distributions from naive sequences to}
\usage{
compareDistancesFromNaiveToMature(dat_a, dat_b)
}
\arguments{
\item{dat_a}{First dataset}

\item{dat_b}{Second dataset}
}
\value{
JS divergence of the two distance distributions
}
\description{
Compare Levenshtein distance distributions from naive sequences to
}
