% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareGCContentDistributions}
\alias{compareGCContentDistributions}
\title{Compare the GC distributions of two lists of DNA sequences}
\usage{
compareGCContentDistributions(dat_a, dat_b,
  column = "sequence_alignment")
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}
}
\value{
JS divergence of the GC content distributions inferred from list_a
  and list_b
}
\description{
Compare the GC distributions of two lists of DNA sequences
}
