% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareGCContents}
\alias{compareGCContents}
\title{Compare the GC distributions of two lists of DNA sequences}
\usage{
compareGCContents(dat_a, dat_b)
}
\arguments{
\item{list_a}{First list or vector of DNA sequences}

\item{list_b}{Second list or vector of DNA sequences}
}
\value{
JS divergence of the GC content distributions inferred from list_a
  and list_b
}
\description{
Compare the GC distributions of two lists of DNA sequences
}
