% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareGeneUsage}
\alias{compareGeneUsage}
\title{Compare gene usage counts for two lists of genes}
\usage{
compareGeneUsage(gene_list_a, gene_list_b, collapse_alleles)
}
\arguments{
\item{gene_list_a}{First vector of genes}

\item{gene_list_b}{Second vector of genes}

\item{collapse_alleles}{Should allelic variants be ignored?}
}
\value{
Mean absolute difference of gene counts between \code{gene_list_a}
  and \code{gene_list_b}
}
\description{
Compare gene usage counts for two lists of genes
}
