% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareGermlineGeneDistributions}
\alias{compareGermlineGeneDistributions}
\title{Compare germline V, D, or J gene usage for two repertoires}
\usage{
compareGermlineGeneDistributions(dat_a, dat_b, gene_type, collapse_alleles)
}
\arguments{
\item{dat_a}{First annotated dataset}

\item{dat_b}{Second annotated dataset}

\item{gene_type}{String for gene type, taken as a column name of the 
annotated datasets. Must be "v_gene", "d_gene", or "j_gene"}

\item{collapse_alleles}{Should allelic variants be ignored?}
}
\value{
Mean absolute difference of gene counts between the two
  repertoires
}
\description{
Compare germline V, D, or J gene usage for two repertoires
}
