% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareGRAVYDistributions}
\alias{compareGRAVYDistributions}
\title{Compare the GRAVY distributions of two datasets}
\usage{
compareGRAVYDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a}{First dataset, a data.table or data.frame object}

\item{dat_b}{Second dataset, a data.table or data.frame object}
}
\value{
The JS divergence of GRAVY distributions
}
\description{
Compare the GRAVY distributions of two datasets
}
