% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareHillNumbers}
\alias{compareHillNumbers}
\title{Compare one or multiple Hill numbers of two datasets. The measure of distance
is a sum of absolute differences (or just the absolute difference if there is
only one diversity_order value}
\usage{
compareHillNumbers(dat_a, dat_b, diversity_orders)
}
\arguments{
\item{dat_a}{First dataset to compare}

\item{dat_b}{Second dataset to compare}

\item{diversity_orders}{Scalar- or vector-valued list of parameters to the
Hill diversity index. Can be any real value although nonnegative values are
recommended as biologically meaningful.}
}
\description{
Compare one or multiple Hill numbers of two datasets. The measure of distance
is a sum of absolute differences (or just the absolute difference if there is
only one diversity_order value
}
