% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareJGene5PrimeDeletionLengthDistributions}
\alias{compareJGene5PrimeDeletionLengthDistributions}
\title{Compare the J gene 5' deletion length distributions of two datasets}
\usage{
compareJGene5PrimeDeletionLengthDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a}{A \code{data.table} corresponding to repertoire annotations}

\item{dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\description{
Compare the J gene 5' deletion length distributions of two datasets
}
