% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareJGeneDistributions}
\alias{compareJGeneDistributions}
\title{Compare J gene distribution between two repertoires}
\usage{
compareJGeneDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a}{First annotated dataset}

\item{dat_b}{Second annotated dataset}
}
\value{
Mean absolute difference of J gene counts between the two
  repertoires
}
\description{
Compare J gene distribution between two repertoires
}
