% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareJointGeneDistributions}
\alias{compareJointGeneDistributions}
\title{Compare joint gene distributions of two datasets}
\usage{
compareJointGeneDistributions(dat_a, dat_b, gene_calls, collapse_alleles,
  by_frequency)
}
\arguments{
\item{dat_a}{A \code{data.table} corresponding to repertoire annotations}

\item{dat_b}{A \code{data.table} corresponding to repertoire annotations}

\item{collapse_alleles}{Should allelic variants be ignored?}
}
\description{
Compare joint gene distributions of two datasets
}
