% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareMutabilityModels}
\alias{compareMutabilityModels}
\title{Compare the mutability models of two datasets}
\usage{
compareMutabilityModels(dat_a, dat_b, sub_mod_a = getSubstitutionModel(dat_a),
  sub_mod_b = getSubstitutionModel(dat_b))
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}

\item{sub_mod_a, sub_mod_b}{A shazam substitution model}
}
\value{
ell-1 divergence of mutability model matrices
}
\description{
Compare the mutability models of two datasets
}
