% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareNNDistanceDistribution}
\alias{compareNNDistanceDistribution}
\title{Compare kth nearest neighbor distance distributions of two lists of
  sequences
\code{compareNNDistanceDistribution} computes the JS divergence of
  the kth nearest neighbor distance distribution of two lists of
  DNA sequences}
\usage{
compareNNDistanceDistribution(list_a, list_b, k = 1)
}
\arguments{
\item{list_a}{First list of sequences}

\item{list_b}{Second list of sequences}

\item{k}{The separation depth for the nearest neighbor distances.
k = 1 corresponds to the nearest neighbor, k = 2 corresponds to
the second-nearest neighbor, etc.}
}
\value{
JS divergence of the distributions inferred from list_a
  and list_b
}
\description{
Compare kth nearest neighbor distance distributions of two lists of
  sequences
\code{compareNNDistanceDistribution} computes the JS divergence of
  the kth nearest neighbor distance distribution of two lists of
  DNA sequences
}
