% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareNNDistanceDistributions}
\alias{compareNNDistanceDistributions}
\title{Compare kth nearest neighbor distance distributions of two lists of
  sequences
\code{compareNNDistanceDistribution} computes the JS divergence of
  the kth nearest neighbor distance distribution of two lists of
  DNA sequences}
\usage{
compareNNDistanceDistributions(dat_a, dat_b,
  column = "sequence_alignment", k = 1, approximate = TRUE, ...)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}

\item{k}{The separation depth for the nearest neighbor distances.
k = 1 corresponds to the nearest neighbor, k = 2 corresponds to
the second-nearest neighbor, etc.}
}
\value{
Estimated JS divergence of the distributions inferred from list_a
  and list_b
}
\description{
Compare kth nearest neighbor distance distributions of two lists of
  sequences
\code{compareNNDistanceDistribution} computes the JS divergence of
  the kth nearest neighbor distance distribution of two lists of
  DNA sequences
}
