% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{comparePairwiseDistanceDistributions}
\alias{comparePairwiseDistanceDistributions}
\title{Compare pairwise distance distributions of two lists of sequences}
\usage{
comparePairwiseDistanceDistributions(dat_a, dat_b, column = "sequence",
  approximate = TRUE)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}

\item{approximate}{If TRUE, uses approximate pairwise distance distributions}
}
\value{
Estimated JS divergence of the distributions inferred from list_a
  and list_b
}
\description{
\code{comparePairwiseDistanceDistributions} computes the JS
  divergence of the pairwise distance distribution of two lists of
  DNA sequences. This function iterates through a number of trials (given
  by \code{trial_count}, subsampling the full datasets by the amount given
  by \code{subsample_count}, and returns the mean divergence.
}
