% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{comparePerGeneMutationRates}
\alias{comparePerGeneMutationRates}
\title{Compare per gene mutation rates of two datasets}
\usage{
comparePerGeneMutationRates(rate_dat_a, rate_dat_b)
}
\arguments{
\item{rate_dat_a, rate_dat_b}{A \code{mutation_rates} object}
}
\value{
l1 divergence of per gene mutation rates
}
\description{
Compare per gene mutation rates of two datasets
}
