% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{comparePerGenePerPositionMutationRates}
\alias{comparePerGenePerPositionMutationRates}
\title{Compare per gene per position mutation rates of two datasets}
\usage{
comparePerGenePerPositionMutationRates(rate_dat_a, rate_dat_b)
}
\arguments{
\item{rate_dat_a}{A \code{mutation_rates} object}

\item{rate_dat_b}{A \code{mutation_rates} object}
}
\description{
Compare per gene per position mutation rates of two datasets
}
