% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{comparePolarityDistributions}
\alias{comparePolarityDistributions}
\title{Compare the polarity distributions of two datasets}
\usage{
comparePolarityDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\description{
Compare the polarity distributions of two datasets
}
