% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{comparePositionalDistanceBetweenMutationsDistributions}
\alias{comparePositionalDistanceBetweenMutationsDistributions}
\title{Compare the positional distance between mutations distributions of two
  datasets}
\usage{
comparePositionalDistanceBetweenMutationsDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\description{
Compare the positional distance between mutations distributions of two
  datasets
}
