% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare.R
\name{compareRepertoires}
\alias{compareRepertoires}
\title{Run through a full repertoire comparison}
\usage{
compareRepertoires(repertoire_1, repertoire_2)
}
\arguments{
\item{repertoire_1}{First repertoire}

\item{repertoire_2}{Second repertoire}
}
\description{
\code{compareRepertoires} iterates through the various comparison functions
  in \code{sumrep}, printing the distance or divergence value of each one.
  Both repertoires are assumed to be both annotated and partitioned.
}
