% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareSackinIndices}
\alias{compareSackinIndices}
\title{Compare Sackin's indices for two phylogenetic trees}
\usage{
compareSackinIndices(tree_1, tree_2)
}
\arguments{
\item{tree_1}{The first phylo object}

\item{tree_2}{The second phylo object}
}
\value{
The absolute difference of the respective Sackin's indices
}
\description{
Compare Sackin's indices for two phylogenetic trees
}
