% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareSelectionEstimates}
\alias{compareSelectionEstimates}
\title{Compare selection estimates for two datasets}
\usage{
compareSelectionEstimates(dat_a, dat_b)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\value{
The estimated JS divergence of selection strength distributions
}
\description{
Compare selection estimates for two datasets
}
