% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareSubstitutionAndMutabilityModels}
\alias{compareSubstitutionAndMutabilityModels}
\title{Compare the substitution and mutability models of two datasets}
\usage{
compareSubstitutionAndMutabilityModels(dat_a, dat_b)
}
\arguments{
\item{dat_a}{First dataset}

\item{dat_b}{Second dataset}
}
\value{
List of two divergences: one for the substitution models and 
  another for the mutability models
}
\description{
Compare the substitution and mutability models of two datasets
}
