% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareVDJDistributions}
\alias{compareVDJDistributions}
\title{Compare joint V, D, and J gene usage between two annotated repertoires}
\usage{
compareVDJDistributions(dat_a, dat_b, collapse_alleles = TRUE)
}
\arguments{
\item{dat_a}{A \code{data.table} corresponding to repertoire annotations}

\item{dat_b}{A \code{data.table} corresponding to repertoire annotations}

\item{collapse_alleles}{Should allelic variants be ignored?}
}
\value{
l1 divergence of counts of VDJ triples between the two repertoires
}
\description{
Compare joint V, D, and J gene usage between two annotated repertoires
}
