% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareVGeneDistributions}
\alias{compareVGeneDistributions}
\title{Compare V gene distribution between two repertoires}
\usage{
compareVGeneDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a}{First annotated dataset}

\item{dat_b}{Second annotated dataset}
}
\value{
Mean absolute difference of V gene counts between the two
  repertoires
}
\description{
Compare V gene distribution between two repertoires
}
