% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{convertDNAToAminoAcids}
\alias{convertDNAToAminoAcids}
\title{Convert a DNA string to a string of single-letter amino acid codes}
\usage{
convertDNAToAminoAcids(sequence)
}
\arguments{
\item{sequence}{String of DNA bases}
}
\value{
String of single-letter amino acid codes
}
\description{
Convert a DNA string to a string of single-letter amino acid codes
}
