% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{convertNucleobasesToAminoAcids}
\alias{convertNucleobasesToAminoAcids}
\title{Convert a vector of DNA strings to a strings of single-letter amino acid codes.
If the string does not correspond to a valid amino acid, convert it to NA.}
\usage{
convertNucleobasesToAminoAcids(sequence_list)
}
\arguments{
\item{sequence_list}{Vector of string of DNA bases}
}
\value{
Vector of strings of single-letter amino acid codes
}
\description{
Convert a vector of DNA strings to a strings of single-letter amino acid codes.
If the string does not correspond to a valid amino acid, convert it to NA.
}
