% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{convertNucleobasesToAminoAcidsBySequence}
\alias{convertNucleobasesToAminoAcidsBySequence}
\title{Convert a single DNA sequence string to an amino acid sequence string}
\usage{
convertNucleobasesToAminoAcidsBySequence(sequence)
}
\arguments{
\item{sequence}{String of DNA bases}
}
\value{
String of single-letter amino acid codes
}
\description{
Convert a single DNA sequence string to an amino acid sequence string
}
