% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approximation.R
\name{doNNSubsamplingBatchStep}
\alias{doNNSubsamplingBatchStep}
\title{Compute the nearest neighbor distribution of a sample of \code{column} of 
  size \code{batch_size} to the full set of sequences in \code{column}.}
\usage{
doNNSubsamplingBatchStep(dat, batch_size = 10, column)
}
\arguments{
\item{dat}{}

\item{batch_size}{The number of sequences sampled during each iteration.}

\item{column}{The column over which the nearest neighbor distance
should be computed. Must correspond to sequence strings, e.g. DNA or
amino acid sequences.}
}
\value{
Sample of \code{batch_size} nearest neighbor distances from 
  \code{dat[[column]]}
}
\description{
Compute the nearest neighbor distribution of a sample of \code{column} of 
  size \code{batch_size} to the full set of sequences in \code{column}.
}
