% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{extractCDR3CodonStartPositions}
\alias{extractCDR3CodonStartPositions}
\title{Extract CDR3 codon start positions from partis-returned dictionary strings}
\usage{
extractCDR3CodonStartPositions(dictionary_list)
}
\arguments{
\item{dictionary_list}{A vector of strings corresponding to a Python 
dictionary, present in the column \code{codon_positions} of a data.table 
returned by \code{annotateSequences}.}
}
\value{
A vector of positions of the CDR3 start codons of the BCR sequences
}
\description{
This is a helper function to \link{getCDR3s}.
Note: We add one to \code{positions} because partis returns position values
  that are zero-based (an artifact) of Python, whereas R is one-based by
  default.
}
