% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{filterAminoAcidSequences}
\alias{filterAminoAcidSequences}
\title{Filter amino acid sequences by removing any 'X' (unrecognized) values,
as well as converting sequences with stop codons to NA}
\usage{
filterAminoAcidSequences(aa_sequences)
}
\arguments{
\item{aa_sequences}{Vector or list of amino acid sequences}
}
\value{
Filtered list with "bad" sequences removed
}
\description{
Filter amino acid sequences by removing any 'X' (unrecognized) values,
as well as converting sequences with stop codons to NA
}
