% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getAliphaticIndex}
\alias{getAliphaticIndex}
\title{Get the aliphatic index of a DNA sequence}
\usage{
getAliphaticIndex(aa_sequence)
}
\arguments{
\item{aa_sequence}{An amino acid sequence string}
}
\value{
The aliphatic index of \code{dna_sequence}, if applicable
}
\description{
Get the aliphatic index of a DNA sequence
}
