% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getAliphaticIndexDistribution}
\alias{getAliphaticIndexDistribution}
\title{Get the distribution of aliphatic indices of a list of sequences}
\usage{
getAliphaticIndexDistribution(sequence_list)
}
\arguments{
\item{sequence_list}{List or vector of DNA sequences}
}
\value{
Vector of aliphatic indices
}
\description{
Get the distribution of aliphatic indices of a list of sequences
}
