% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getAminoAcidDistribution}
\alias{getAminoAcidDistribution}
\title{Compute a table of counts for each amino acid in the list of sequences}
\usage{
getAminoAcidDistribution(dat, column = "junction_aa", standardize = TRUE)
}
\arguments{
\item{sequences}{Vector of amino acid sequences}
}
\value{
table of counts for each amino acid in \code{sequences}
}
\description{
Compute a table of counts for each amino acid in the list of sequences
}
