% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getAminoAcidProperties}
\alias{getAminoAcidProperties}
\title{Get a particular distribution of amino acid properties using the
  \code{alakazam::aminoAcidProperties} function}
\usage{
getAminoAcidProperties(dat, column = "junction_aa", suffix)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}

\item{suffix}{A property-specific suffix used to extract the desired
property from the \code{data.table} returned by \code{alakazam}. 
This is usually of the form "AA_PROPERTY" (e.g. "AA_POLARITY"),
but not always (e.g. "AA_BASIC" for basicity)}
}
\value{
distribution of the given amino acid property
}
\description{
Get a particular distribution of amino acid properties using the
  \code{alakazam::aminoAcidProperties} function
}
