% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getApproximateDistribution}
\alias{getApproximateDistribution}
\title{Approximate the distribution of \code{summary_function} for \code{dat} by 
  \code{summary_function} for the subsample, and accumulating a representative
  distribution. The function stops when \code{divergence_function} between two
  distribution iterates is sufficiently small.}
\usage{
getApproximateDistribution(dat, summary_function, sample_count = 100,
  tol = 0.001, divergence_function = getContinuousJSDivergence)
}
\arguments{
\item{dat}{The dataset for which the distribution is desired}

\item{summary_function}{A function which computes a summary statistic of \code{dat},
which yields a distribution}

\item{sample_count}{The size of the subsample for each iteration}

\item{tol}{The threshold for \code{divergence_function} to be considered converged}

\item{divergence_function}{The divergence computed between successive iterates.
This should probably be the JS divergence, specified for discrete or continuous data
depending on the nature of \code{summary_function}}
}
\description{
Approximate the distribution of \code{summary_function} for \code{dat} by 
  \code{summary_function} for the subsample, and accumulating a representative
  distribution. The function stops when \code{divergence_function} between two
  distribution iterates is sufficiently small.
}
