% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approximation.R
\name{getApproximateNearestNeighborDistribution}
\alias{getApproximateNearestNeighborDistribution}
\title{Runs a special approximation routine since nearest neighbor distances are
  not a regular distribution for subsampling. (Since the nearest neighbor
  distance of a sample is necessarily larger than that of the population)
  So, we cluster the sequences using vsearch and sample sequences based
  on cluster frequency. This ensures no nearest neighbor distance is higher
  than the given cutoff.}
\usage{
getApproximateNearestNeighborDistribution(dat, column, k = 1,
  pairwise_identity = 0, tol = 0.01, batch_size = 100)
}
\description{
Runs a special approximation routine since nearest neighbor distances are
  not a regular distribution for subsampling. (Since the nearest neighbor
  distance of a sample is necessarily larger than that of the population)
  So, we cluster the sequences using vsearch and sample sequences based
  on cluster frequency. This ensures no nearest neighbor distance is higher
  than the given cutoff.
}
