% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getAtchleyFactorDistribution}
\alias{getAtchleyFactorDistribution}
\title{Get the distribution of a given Atchley factor for a vector of amino acid 
  sequences}
\usage{
getAtchleyFactorDistribution(aa_sequences, factor_number)
}
\arguments{
\item{aa_sequences}{A vector of amino acid sequence string}

\item{factor_number}{The Atchley factor to be applied over 
\code{aa_sequences}. Must be 1, 2, 3, 4, or 5.}
}
\value{
A vector of Atchley factors of each amino acid in each sequence
}
\description{
Get the distribution of a given Atchley factor for a vector of amino acid 
  sequences
}
