% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getAtchleyFactorDistributions}
\alias{getAtchleyFactorDistributions}
\title{Get a list of the distributions of each of the five Atchley factors for a 
  dataset.}
\usage{
getAtchleyFactorDistributions(dat, column = "junction_aa")
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}
}
\value{
List of Atchley factor distributions
}
\description{
Get a list of the distributions of each of the five Atchley factors for a 
  dataset.
}
