% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getAtchleyFactorList}
\alias{getAtchleyFactorList}
\title{Get the mean Atchley factor of an amino acid sequence}
\usage{
getAtchleyFactorList(aa_seq, factor_number)
}
\arguments{
\item{aa_seq}{String of an amino acid sequence}

\item{factor_number}{The Atchley factor to be applied to \code{aa_seq}. 
Must be 1, 2, 3, 4, or 5.}
}
\value{
The mean of the set of Atchley factors for 
  \code{aa_seq}
}
\description{
Get the mean Atchley factor of an amino acid sequence
}
