% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivergenceFunctions.R
\name{getAutomaticAverageDivergence}
\alias{getAutomaticAverageDivergence}
\title{Estimate the divergence by subsampling and averaging until a stable estimate
  is attained}
\usage{
getAutomaticAverageDivergence(dataset_a, dataset_b, summary_function,
  subsample_count, divergenceFunction = getJSDivergence,
  tolerance = 0.01, ...)
}
\arguments{
\item{dataset_a}{First dataset (vector, matrix, or data.table) input to 
\code{func}}

\item{dataset_b}{Second dataset input to \code{func}}

\item{summary_function}{Comparison function for which to calculate the JS 
divergence}

\item{subsample_count}{Number of samples to extract on each trial}

\item{divergenceFunction}{The divergence function to apply to the two 
datasets. JS divergence by default, although this only makes sense for
vectors corresponding to empirical distributions}

\item{tolerance}{The criterion to determine convergence based on relative
difference between iterates}
}
\value{
The average computed divergence across trials, used as an
  estimate of the true JS divergence
}
\description{
Estimate the divergence by subsampling and averaging until a stable estimate
  is attained
}
