% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getAverageDivergence}
\alias{getAverageDivergence}
\title{Estimate a divergence by subsampling and averaging over the comparison
  function output}
\usage{
getAverageDivergence(dataset_a, dataset_b, func, subsample_count, trial_count,
  divergenceFunction = getJSDivergence, ...)
}
\arguments{
\item{dataset_a}{First dataset (vector, matrix, or data.table) input to 
\code{func}}

\item{dataset_b}{Second dataset input to \code{func}}

\item{func}{Comparison function for which to calculate the JS divergence}

\item{subsample_count}{Number of samples to extract on each trial}

\item{trial_count}{Number of trials to average over}

\item{divergenceFunction}{The divergence function to apply to the two 
datasets. JS divergence by default, although this only makes sense for
vectors corresponding to empirical distributions}
}
\value{
The average computed JS divergence across trials, used as an
  estimate of the true JS divergence
}
\description{
Arguemnts to \code{func} may be supplied after the other required 
  parameters.
}
\examples{
seq_1 <- c("AAAAA", "AATTT", "TTATT", "ACATG", "GGGAA")
seq_2 <- c("AAAAC", "ACATG", "CGGGA", "ACATG", "GGACA")
getAverageDivergence(seq_1, seq_2, getNearestNeighborDistances, 
  subsample_count=3, trial_count=20, k=2)
}
