% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getAverageJSDivergence}
\alias{getAverageJSDivergence}
\title{Estimate JS divergence by subsampling and averaging over the comparison
  function output}
\usage{
getAverageJSDivergence(sequence_a, sequence_b, func, subsample_count,
  trial_count, ...)
}
\arguments{
\item{sequence_a}{First sequence input to \code{func}}

\item{sequence_b}{Second sequence input to \code{func}}

\item{func}{Comparison function for which to calculate the JS divergence}

\item{subsample_count}{Number of samples to extract on each trial}

\item{trial_count}{Number of trials to average over}
}
\value{
The average computed JS divergence across trials, used as an
  estimate of the true JS divergence
}
\description{
Arguemnts to \code{func} may be supplied after the other required 
  parameters.
}
\examples{
seq_1 <- c("AAAAA", "AATTT", "TTATT", "ACATG", "GGGAA")
seq_2 <- c("AAAAC", "ACATG", "CGGGA", "ACATG", "GGACA")
getAverageJSDivergence(seq_1, seq_2, getNearestNeighborDistances, 
  subsample_count=3, trial_count=20, k=2)
}
