% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getBasicityDistribution}
\alias{getBasicityDistribution}
\title{Get a vector of basicity values for a set of an amino acid sequences}
\usage{
getBasicityDistribution(dat, column = "junction_aa")
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}
}
\description{
Get a vector of basicity values for a set of an amino acid sequences
}
