% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getCDR3LengthDistribution}
\alias{getCDR3LengthDistribution}
\title{Get the distribution of inferred length of CDR3 regions of each sequence.
  Requires either the \code{cdr3_length} or \code{junction} column for
  \code{by_amino_acid=FALSE}, or the \code{junction_aa} column for
  \code{by_amnio_acid=TRUE}.}
\usage{
getCDR3LengthDistribution(dat, by_amino_acid = FALSE)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{by_amino_acid}{If TRUE, the length is computed in terms of amino 
acids; otherwise, the length is computed in terms of nucleotides.}
}
\value{
Vector of CDR3 lengths (in nt units)
}
\description{
Get the distribution of inferred length of CDR3 regions of each sequence.
  Requires either the \code{cdr3_length} or \code{junction} column for
  \code{by_amino_acid=FALSE}, or the \code{junction_aa} column for
  \code{by_amnio_acid=TRUE}.
}
