% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getCDR3Lengths}
\alias{getCDR3Lengths}
\title{Get the distribution of inferred length of CDR3 regions of each sequence}
\usage{
getCDR3Lengths(dat)
}
\arguments{
\item{Annotated}{dataset}
}
\value{
Vector of CDR3 lengths (in nt units)
}
\description{
Get the distribution of inferred length of CDR3 regions of each sequence
}
