% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getCDR3s}
\alias{getCDR3s}
\title{Get a vector of CDR3 DNA strings of an annotated dataset}
\usage{
getCDR3s(dat)
}
\arguments{
\item{dat}{Dataset, either a data.table or data.frame object}
}
\value{
Vector of CDR3 strings
}
\description{
Warning: This function is called on the raw dataset from 
  \code{doFullAnnotation} before \code{input_seqs} is changed to
  \code{mature_seq}.
}
