% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getColdspotCount}
\alias{getColdspotCount}
\title{Get the number of occurrences of AID coldspots in a set of reference 
  sequences}
\usage{
getColdspotCount(dna_sequences, coldspots = c("SYC"))
}
\value{
The number of AID coldhotspot occurrences in \code{dna_sequences}
}
\description{
Get the number of occurrences of AID coldspots in a set of reference 
  sequences
}
