% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getColdspotCountDistribution}
\alias{getColdspotCountDistribution}
\title{Get full or approximate distribution of coldspot counts for a column of
  a given dataset}
\usage{
getColdspotCountDistribution(dat, column = "sequence_alignment",
  coldspots = "SYC", approximate = FALSE, ...)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}

\item{coldspots}{Vector of coldspots of interest}

\item{approximate}{If TRUE, approximate distribution by subsampling.}
}
\description{
Get full or approximate distribution of coldspot counts for a column of
  a given dataset
}
