% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getCollessLikeIndex}
\alias{getCollessLikeIndex}
\title{Compute Colless-like index for the given tree}
\usage{
getCollessLikeIndex(tree)
}
\arguments{
\item{tree}{A phylo object corresponding to a phylogeny}
}
\value{
Colless-like index of tree balance
}
\description{
Compute Colless-like index for the given tree
}
