% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getColumnSequences}
\alias{getColumnSequences}
\title{This function attempts to access the provided column from the provided
  dat, perfoming any optional filters that were specified, and returns
  a vector of the relevant sequence strings.}
\usage{
getColumnSequences(dat, column, drop_gaps, remove_stop_codon_seqs = TRUE,
  in_frame_only = TRUE)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{A string naming the expected column of \code{dat}}

\item{drop_gaps}{If TRUE, removes '.' and '-' gap characters from each
sequence}

\item{remove_stop_codon_seqs}{If TRUE, removes any sequence with a 
stop codon. This assumes the \code{stop_codon} column is present in
\code{dat}, as defined by the AIRR rearrangement schema.}

\item{in_frame_only}{If TRUE, removes any sequence with an
out-of-frame V or J segment.}
}
\value{
A vector of filtered strings from by \code{dat[[column]]}, 
  if available
}
\description{
This function attempts to access the provided column from the provided
  dat, perfoming any optional filters that were specified, and returns
  a vector of the relevant sequence strings.
}
