% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getColumnValues}
\alias{getColumnValues}
\title{A wrapper of \code{alakazam::checkColumns} that stops execution when an
  error is thrown, and returns the vector of column values if available.}
\usage{
getColumnValues(dat, column)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{A string naming the expected column of \code{dat}}
}
\value{
A vector of values given by \code{dat[[column]]}, if available
}
\description{
A wrapper of \code{alakazam::checkColumns} that stops execution when an
  error is thrown, and returns the vector of column values if available.
}
