% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getContinuousJSDivergence}
\alias{getContinuousJSDivergence}
\title{Compute JS divergence for continuous samples}
\usage{
getContinuousJSDivergence(sample_1, sample_2)
}
\arguments{
\item{sample_1}{First data sample}

\item{sample_2}{Second data sample}
}
\value{
Approximate JS divergence of the distributions induced from sample_1
  and sample_2
}
\description{
This is currently not the default method for continuous data, as it requires
  constructing approximating functions of the two densities, and then using
  numerical integrals for the expectation steps.
}
