% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getCopheneticIndex}
\alias{getCopheneticIndex}
\title{Compute cophenetic index for the given tree}
\usage{
getCopheneticIndex(tree)
}
\arguments{
\item{tree}{A phylo object corresponding to a phylogeny}
}
\value{
cophenetic index of tree balance
}
\description{
Compute cophenetic index for the given tree
}
