% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getDeletionLengths}
\alias{getDeletionLengths}
\title{Get the requested deletion lengths of a dataset}
\usage{
getDeletionLengths(dat, column)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the desired deletion
lengths.}
}
\value{
Vector of deletion lengths
}
\description{
Get the requested deletion lengths of a dataset
}
