% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getDGene3PrimeDeletionLengthDistribution}
\alias{getDGene3PrimeDeletionLengthDistribution}
\title{Get the distribution of D gene 3' deletion lengths of a dataset}
\usage{
getDGene3PrimeDeletionLengthDistribution(dat)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}
}
\value{
Distribution of D gene 3' deletion lengths
}
\description{
Get the distribution of D gene 3' deletion lengths of a dataset
}
