% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getDistanceFromGermlineToSequenceDistribution}
\alias{getDistanceFromGermlineToSequenceDistribution}
\title{Get the exact or approximate distribution of Levenshtein distances from 
  the inferred naive sequences to the observed mature ones}
\usage{
getDistanceFromGermlineToSequenceDistribution(dat, approximate = FALSE,
  v_gene_only = FALSE)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{approximate}{If TRUE, approximate distribution by subsampling.}

\item{v_gene_only}{If TRUE, restrict sequences to the V gene only}
}
\value{
Vector of Levenshtein distances from naive to mature
}
\description{
Get the exact or approximate distribution of Levenshtein distances from 
  the inferred naive sequences to the observed mature ones
}
