% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getDistanceMatrix}
\alias{getDistanceMatrix}
\title{Get distance matrix from stringdist}
\usage{
getDistanceMatrix(raw_sequences)
}
\arguments{
\item{raw_sequences}{List or vector of DNA sequences}
}
\value{
Distance matrix of the sequences, using hamming distances if all
  sequences are the same length, and levenshtein otherwise
}
\description{
\code{getDistanceMatrix} determines
}
