% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getDistancesFromNaiveToMature}
\alias{getDistancesFromNaiveToMature}
\title{Get the exact distribution of Levenshtein distances from the inferred 
  naive sequences to the observed mature ones, sequence by sequence}
\usage{
getDistancesFromNaiveToMature(dat, v_gene_only = FALSE)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{v_gene_only}{If TRUE, restrict sequences to the V gene only}
}
\value{
Vector of Levenshtein distances from naive to mature
}
\description{
Get the exact distribution of Levenshtein distances from the inferred 
  naive sequences to the observed mature ones, sequence by sequence
}
