% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getDistanceVector}
\alias{getDistanceVector}
\title{Get sorted vector of pairwise distances}
\usage{
getDistanceVector(sequence_list)
}
\arguments{
\item{sequence_list}{vector of sequence strings (DNA, AA, etc.)}
}
\value{
Vector of pairwise distances
}
\description{
Convert the matrix given by \link{getDistanceMatrix} into a sorted
  vector of distance values.
}
