% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartisFunctions.R
\name{getFullPartisAnnotation}
\alias{getFullPartisAnnotation}
\title{Do extended partis processing of annotations}
\usage{
getFullPartisAnnotation(output_path, output_file, partis_path)
}
\arguments{
\item{output_path}{Desired output directory, which will contain 
\code{output_filename}}

\item{partis_path}{The full path to the partis executable}
}
\description{
Do extended partis processing of annotations
}
