% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getGCContent}
\alias{getGCContent}
\title{Get the GC content of each sequence in a vector of strings}
\usage{
getGCContent(raw_sequences)
}
\arguments{
\item{raw_sequences}{Vector of sequence strings}
}
\value{
Vector of GC content values
}
\description{
Get the GC content of each sequence in a vector of strings
}
