% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getGCContentDistribution}
\alias{getGCContentDistribution}
\title{Get the GC content distribution of a list of DNA sequences}
\usage{
getGCContentDistribution(raw_sequences)
}
\arguments{
\item{raw_sequences}{List or vector of strings or character sequences
corresponding to DNA sequences}
}
\value{
A vector of GC content values
}
\description{
\code{getGCContentDistribution} returns a list of the GC content of each DNA
  sequence in \code{raw_sequences}, given by the \code{ape} library
}
