% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getGeneUsageTableFromFullList}
\alias{getGeneUsageTableFromFullList}
\title{Get gene usage table from full list of genes, setting zero to unused ones}
\usage{
getGeneUsageTableFromFullList(gene_list, full_gene_list)
}
\arguments{
\item{gene_list}{List of genes being used in a given repertoire}

\item{gene_list}{List of all available genes present in at least one
repertoire under consideration}
}
\description{
Get gene usage table from full list of genes, setting zero to unused ones
}
