% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getGridDims}
\alias{getGridDims}
\title{Handy automatic dimension retrieval given number of grid cells}
\usage{
getGridDims(n)
}
\description{
Handy automatic dimension retrieval given number of grid cells
}
