% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getHotspotCount}
\alias{getHotspotCount}
\title{Get the number of occurrences of AID hotspots in a set of reference 
  sequences}
\usage{
getHotspotCount(dat, column, hotspots)
}
\value{
The number of AID hotspot occurrences in \code{dna_sequences}
}
\description{
Get the number of occurrences of AID hotspots in a set of reference 
  sequences
}
