% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getHotspotCountDistribution}
\alias{getHotspotCountDistribution}
\title{Get the distribution of hotspot counts of sequences in column \code{column}
  of \code{dat}}
\usage{
getHotspotCountDistribution(dat, column = "mature_seq", hotspots = c("WRC",
  "WA"), approximate = TRUE)
}
\description{
Get the distribution of hotspot counts of sequences in column \code{column}
  of \code{dat}
}
