% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getHydrophobicityFromAASequence}
\alias{getHydrophobicityFromAASequence}
\title{Get hydrophobicity from amino acid sequence, corresponding to
the fourth kidera factor}
\usage{
getHydrophobicityFromAASequence(sequence)
}
\arguments{
\item{sequence}{Amino acid sequence string}
}
\value{
real-valued hydrophobicity estimate, computed as the average
  hydrophobicity of each amino acid in the sequence
}
\description{
Get hydrophobicity from amino acid sequence, corresponding to
the fourth kidera factor
}
