% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgBlastFunctions.R
\name{getIgBlastAnnotations}
\alias{getIgBlastAnnotations}
\title{Get IgBLAST annotations from an input fasta file}
\usage{
getIgBlastAnnotations(input_filename, output_filename = "igblast_out",
  locus = "igh", organism = "human", domain_system = "imgt",
  ig_seqtype = "Ig", num_threads = 8, igblast_dir, changeo_dir)
}
\arguments{
\item{input_filename}{The path to the input fasta file}

\item{output_filename}{The desired output filename WITHOUT an extension}

\item{locus}{The locus from which the sequences are derived}

\item{organism}{The organism/species to which the sequences belong}

\item{domain_system}{(input to IgBLAST)}

\item{ig_seqtype}{(input to IgBLAST}
}
\description{
Currently kind of hacky, as it changes directories to the location of
  igblast and its germline databases, runs some commands, and changes back
  to the initial directory.
  Change-O is used to parse the raw output of igblastn.
  TODO: Delete interim files, try-catch the success of changeo
}
