% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getInFramePercentage}
\alias{getInFramePercentage}
\title{Get the percentage of sequences in a dataset which, after VDJ recombination 
  and SHM, have the start of the CDR3 in line with the start of the germline 
  V sequence}
\usage{
getInFramePercentage(dat, column = "vj_in_frame")
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}
}
\value{
The percentage of in-frame sequences in \code{dat}
}
\description{
Get the percentage of sequences in a dataset which, after VDJ recombination 
  and SHM, have the start of the CDR3 in line with the start of the germline 
  V sequence
}
