% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getInsertionLengths}
\alias{getInsertionLengths}
\title{Get the insertion lengths corresponding to a column of a given datset}
\usage{
getInsertionLengths(dat, column)
}
\arguments{
\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}
}
\description{
Get the insertion lengths corresponding to a column of a given datset
}
