% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getInsertionMatrix}
\alias{getInsertionMatrix}
\title{Get the Markov transition matrix for either VD or DJ insertions}
\usage{
getInsertionMatrix(dat, column)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{The name of the column corresponding to the inserted sequences}
}
\value{
The empirical transition matrix for each (base, base) pair.
}
\description{
Get the Markov transition matrix for either VD or DJ insertions
}
