% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getJGene5PrimeDeletionLengthDistribution}
\alias{getJGene5PrimeDeletionLengthDistribution}
\title{Get the distribution of J gene 5' deletion lengths of a dataset}
\usage{
getJGene5PrimeDeletionLengthDistribution(dat)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}
}
\value{
Distribution of J gene 5' deletion lengths
}
\description{
Get the distribution of J gene 5' deletion lengths of a dataset
}
