% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getJointGeneTable}
\alias{getJointGeneTable}
\title{Get table of combined V, D, and J usage frequencies}
\usage{
getJointGeneTable(dat, gene_calls = c("v_call", "j_call"),
  collapse_alleles = TRUE, by_frequency = TRUE)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{collapse_alleles}{Should allelic variants be ignored?}

\item{by_frequency}{If true, scale table entries as relative frequencies.
Otherwise, entries will be counts.}
}
\value{
A table of joint gene IDs and usage counts
}
\description{
Get table of combined V, D, and J usage frequencies
}
