% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getKideraFactorDistributions}
\alias{getKideraFactorDistributions}
\title{Get distributions for each of the ten Kidera factors for a dataset}
\usage{
getKideraFactorDistributions(dat, column = "junction_aa", as_list = TRUE)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}

\item{as_list}{If TRUE, return a list of the ten distributions. Otherwise,
return an (n)x(10) data.table, where n is the number of valid amino acid 
sequences in \code{column}}
}
\value{
List or data.table of Kidera factor distributions
}
\description{
Get distributions for each of the ten Kidera factors for a dataset
}
