% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getKideraFactorsBySequence}
\alias{getKideraFactorsBySequence}
\title{Get the kidera factors of a single sequence.
  See the \code{Peptides::kideraFactors} documentation for more information.}
\usage{
getKideraFactorsBySequence(aa_sequence)
}
\arguments{
\item{aa_sequence}{An amino acid sequence string}
}
\value{
A named vector of the 10 Kidera factors for \code{aa_sequence}
}
\description{
Get the kidera factors of a single sequence.
  See the \code{Peptides::kideraFactors} documentation for more information.
}
