% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgBlastFunctions.R
\name{getMatureSequences}
\alias{getMatureSequences}
\title{Add n's to positions for which the input (mature) sequence did not contain
  a base, but for which the inferred naive (germline) sequence does contain
  a base. IgBlast/Change-O by default place '.'s in these positions, which
  is also the symbol to denote IMGT gaps. Thus, manual processing to account
  for this is required.}
\usage{
getMatureSequences(germline_imgt, sequence_imgt)
}
\description{
Add n's to positions for which the input (mature) sequence did not contain
  a base, but for which the inferred naive (germline) sequence does contain
  a base. IgBlast/Change-O by default place '.'s in these positions, which
  is also the symbol to denote IMGT gaps. Thus, manual processing to account
  for this is required.
}
