% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivergenceFunctions.R
\name{getMeanAbsoluteDifference}
\alias{getMeanAbsoluteDifference}
\title{Get the mean absolute difference of two sequences}
\usage{
getMeanAbsoluteDifference(sequence_a, sequence_b, ignore_na = TRUE)
}
\arguments{
\item{sequence_a}{First sequence, either a vector or matrix}

\item{sequence_b}{Second sequence}

\item{ignore_na}{Should we ignore na values when computing the mean?}
}
\description{
Get the mean absolute difference of two sequences
}
