% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getMeanAtchleyFactorBySequence}
\alias{getMeanAtchleyFactorBySequence}
\title{Get the mean Atchley factor of a DNA sequence}
\usage{
getMeanAtchleyFactorBySequence(dna_seq)
}
\arguments{
\item{dna_seq}{String of a DNA sequence}
}
\value{
The mean of the set of Atchley factors for 
  \code{dna_seq}
}
\description{
TODO: Do something more clever than taking the mean. Is there
  a set of factors of interest that we can isolate? Or should
  we do an average pairwise comparison?
}
