% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getMotifCount}
\alias{getMotifCount}
\title{Get number of times a motif occurs in a set of one or more reference 
  sequences}
\usage{
getMotifCount(motif, dna_sequences)
}
\arguments{
\item{motif}{String representing the motif pattern}

\item{dna_sequences}{List, vector of reference sequences}
}
\value{
The number of occurrences of \code{motif} in \code{dna_sequences}
}
\description{
Get number of times a motif occurs in a set of one or more reference 
  sequences
}
