% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getMutabilityModel}
\alias{getMutabilityModel}
\title{Get the inferred mutability model of a repetoire}
\usage{
getMutabilityModel(dat, substitution_model)
}
\arguments{
\item{dat}{Annotated dataset}

\item{substitution_model}{The inferred substitution model, which can be
obtained via \code{getSubstitutionModel}}
}
\value{
The default shazam mutability model
}
\description{
This requires the inferred substitution model of \code{dat} as input
}
