% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartisFunctions.R
\name{getMutationInfo}
\alias{getMutationInfo}
\title{Get per-gene and per-gene-per-position mutation information from partis}
\usage{
getMutationInfo(filename)
}
\arguments{
\item{filename}{YAML file from partis output folder corresponding to a gene}
}
\value{
Object containing the gene's overall mutation rate as well as a 
  vector of positional mutation rates
}
\description{
\code{getMutationInfo} is run if \code{doFullAnnotation} is true in the 
  \link{annotateSequences} call. This function extracts mutation information
  from the given YAML file in the parameter directory output by partis.
}
