% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getNameFromFunctionString}
\alias{getNameFromFunctionString}
\title{Helper function to get abbreviated summary names from getter function names
  or divergence object names}
\usage{
getNameFromFunctionString(function_string)
}
\description{
Helper function to get abbreviated summary names from getter function names
  or divergence object names
}
