% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getPairwiseDistanceDistribution}
\alias{getPairwiseDistanceDistribution}
\title{Get an exact or approximate distribution of pairwise distances}
\usage{
getPairwiseDistanceDistribution(dat, column = "mature_seq",
  approximate = TRUE, ...)
}
\arguments{
\item{approximate}{if TRUE, approximate the distribution by subsampling
and averaging}

\item{sequence_list}{vector of DNA sequence strings}
}
\value{
vector of integer-valued distances
}
\description{
Get an exact or approximate distribution of pairwise distances
}
