% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartisFunctions.R
\name{getPartisAnnotations}
\alias{getPartisAnnotations}
\title{Get partis annotations from the partis output directory.
This function does NOT run partis from R.}
\usage{
getPartisAnnotations(output_path, output_filename,
  partis_path = Sys.getenv("PARTIS_PATH"), do_full_annotation = TRUE,
  collapse_clones = TRUE)
}
\arguments{
\item{output_filename}{The --infname argument to partis}

\item{partis_path}{The path to the partis binary}

\item{do_full_annotation}{Does further processing than the default by partis}

\item{collapse_clones}{Convert a row of colon-separated clonal families to
separate rows for each member}

\item{filepath}{The output directory from the partis call}
}
\description{
Get partis annotations from the partis output directory.
This function does NOT run partis from R.
}
