% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getPerGeneMutationRates}
\alias{getPerGeneMutationRates}
\title{Get mutation rates of each gene from a \code{mutation_rates} object.
  This object is inferred from some XCR repertoire and should be an element 
  of a \code{list}, e.g. as returned by \code{getPartisAnnotations}.}
\usage{
getPerGeneMutationRates(rate_dat)
}
\arguments{
\item{rate_dat}{A \code{mutation_rates} object containing, e.g. as obtained
  in \code{getPartisAnnotations}
return A named vector of mutation rates for each gene}
}
\description{
Get mutation rates of each gene from a \code{mutation_rates} object.
  This object is inferred from some XCR repertoire and should be an element 
  of a \code{list}, e.g. as returned by \code{getPartisAnnotations}.
}
