% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getPerGenePerPositionMutationRates}
\alias{getPerGenePerPositionMutationRates}
\title{Get mutation rates by position of each gene from a \code{mutation_rates}
  object.}
\usage{
getPerGenePerPositionMutationRates(rate_dat)
}
\arguments{
\item{rate_dat}{A \code{mutation_rates} object containing, e.g. as obtained
  in \code{getPartisAnnotations}
return A named vector of mutation rates for each gene}
}
\description{
Get mutation rates by position of each gene from a \code{mutation_rates}
  object.
}
