% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getPositionalDistanceBetweenMutationsDistribution}
\alias{getPositionalDistanceBetweenMutationsDistribution}
\title{Get the distribution of positional distances between mutations of a given
  dataset.
  Note that this requires \code{nchar} to match for each naive/mature 
  sequence pair. (TODO: relax this assumption by discarding cases which
  don't match)}
\usage{
getPositionalDistanceBetweenMutationsDistribution(dat,
  naive_column = "germline_alignment",
  mature_column = "sequence_alignment")
}
\arguments{
\item{dat}{}

\item{naive_column}{Column containing the inferred naive/germline sequences}

\item{mature_column}{Column containing the mature SHM-experienced sequences}

\item{return}{Vector of positional distances}
}
\description{
Get the distribution of positional distances between mutations of a given
  dataset.
  Note that this requires \code{nchar} to match for each naive/mature 
  sequence pair. (TODO: relax this assumption by discarding cases which
  don't match)
}
