% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getPositionalDistancesBetweenMutationsBySequence}
\alias{getPositionalDistancesBetweenMutationsBySequence}
\title{Get the positional distances between mutations given a single mature sequence
  and corresponding naive sequence}
\usage{
getPositionalDistancesBetweenMutationsBySequence(naive, mature)
}
\arguments{
\item{naive}{A string corresponding to a naive sequence}

\item{mature}{A string corresponding to a mature sequence}
}
\value{
A vector of positional distances between the mutations
}
\description{
Get the positional distances between mutations given a single mature sequence
  and corresponding naive sequence
}
