% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getSackinIndex}
\alias{getSackinIndex}
\title{Compute Sackin's index for the given tree}
\usage{
getSackinIndex(tree)
}
\arguments{
\item{tree}{A phylo object corresponding to a phylogeny}
}
\value{
Sackin's index of tree balance
}
\description{
Compute Sackin's index for the given tree
}
