% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getSelectionEstimate}
\alias{getSelectionEstimate}
\title{Get selection estimates via the shazam Baseline models}
\usage{
getSelectionEstimate(dat)
}
\arguments{
\item{dat}{Annotated dataset}
}
\value{
Distribution of baseline values, which are measures of selection 
  strength. Thus, Sigma = 0 corresponds to no selection, Sigma > 0 
  corresponds to positive selection, and Sigma < 0 corresponds to negative
  selection
}
\description{
shazam::summarizeBaseline appends mutliple columns to \code{dat}, including
  BASELINE_SIGMA, which is the estimated selection strength of the 
  repertoire
}
