% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getSequenceListFromFasta}
\alias{getSequenceListFromFasta}
\title{Import a vector of DNA sequence strings from a fasta file}
\usage{
getSequenceListFromFasta(filename)
}
\arguments{
\item{filename}{Name of fasta file including the sequences}
}
\value{
A vector of DNA sequence strings
}
\description{
Import a vector of DNA sequence strings from a fasta file
}
