% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getSpotCount}
\alias{getSpotCount}
\title{Get number of times a set of motif (hot/cold)spots occur in a set of 
  reference sequences}
\usage{
getSpotCount(dna_sequences, spots)
}
\arguments{
\item{spots}{Vector of hot or cold spots of interest}
}
\value{
The total number of occurrences of each motif in \code{spots}, in
  \code{dna_sequences}
}
\description{
Get number of times a set of motif (hot/cold)spots occur in a set of 
  reference sequences
}
