% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getVDInsertionLengthDistribution}
\alias{getVDInsertionLengthDistribution}
\title{Get the distribution of VD insertion lengths of a dataset}
\usage{
getVDInsertionLengthDistribution(dat)
}
\description{
Get the distribution of VD insertion lengths of a dataset
}
