% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getVGene3PrimeDeletionLengthDistribution}
\alias{getVGene3PrimeDeletionLengthDistribution}
\title{Get the distribution of V gene 3' deletion lengths of a dataset}
\usage{
getVGene3PrimeDeletionLengthDistribution(dat)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}
}
\value{
Distribution of V gene 3' deletion lengths
}
\description{
Get the distribution of V gene 3' deletion lengths of a dataset
}
