% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getVJInsertionLengthDistribution}
\alias{getVJInsertionLengthDistribution}
\title{This calls the function for VD insertions since the AIRR field
  for VD and VJ insertion lengths is the same}
\usage{
getVJInsertionLengthDistribution(dat)
}
\description{
This calls the function for VD insertions since the AIRR field
  for VD and VJ insertion lengths is the same
}
