% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{hasStopCodon}
\alias{hasStopCodon}
\title{Check if an amino acid sequence has a stop codon}
\usage{
hasStopCodon(aa_sequence)
}
\arguments{
\item{aa_sequence}{String of single-letter amino acid codes}
}
\value{
A boolean stating whether any stop codons are present 
  in \code{aa_sequence}
}
\description{
Check if an amino acid sequence has a stop codon
}
