% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{hasUnrecognizedAminoAcids}
\alias{hasUnrecognizedAminoAcids}
\title{Check if an amino acid sequence has unrecognized codons}
\usage{
hasUnrecognizedAminoAcids(aa_sequence)
}
\arguments{
\item{aa_sequence}{String of single-letter amino acid codons}
}
\value{
A boolean stating whether any unrecognized AA codons are present 
  in \code{aa_sequence}
}
\description{
Check if an amino acid sequence has unrecognized codons
}
