% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{loadNewDatasets}
\alias{loadNewDatasets}
\title{Load datasets that are not already in the workspace.
  The variable names are taken to be exactly the name of the corresponding
  .rds files}
\usage{
loadNewDatasets(data_dir)
}
\arguments{
\item{data_dir}{The directory which contains .rds files to be loaded}
}
\description{
Load datasets that are not already in the workspace.
  The variable names are taken to be exactly the name of the corresponding
  .rds files
}
