% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{multiplot}
\alias{multiplot}
\title{Display an array of ggplots within one figure}
\usage{
multiplot(plotlist = NULL, cols, rows, layout = NULL,
  tall_plot = FALSE, ...)
}
\arguments{
\item{plotlist}{List of plots created with ggplot}

\item{cols}{Number of columns for the figure}

\item{rows}{Number of rows for the figure}
}
\description{
Display an array of ggplots within one figure
}
