% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{parsePythonDictionary}
\alias{parsePythonDictionary}
\title{Parse a python dictionary string into an R list}
\usage{
parsePythonDictionary(dictionary)
}
\arguments{
\item{dictionary}{String that represents a dictionary in Python's syntax}
}
\value{
An R list containing the same information as \code{dictionary}
}
\description{
Parse a python dictionary string into an R list
}
