% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartisFunctions.R
\name{partitionSequences}
\alias{partitionSequences}
\title{Perform clonal partitioning with partis}
\usage{
partitionSequences(input_filename, output_filename = "partition.csv",
  partis_path = Sys.getenv("PARTIS_PATH"), num_procs = 4, cleanup = TRUE,
  output_path = "_output", germline_dir = NULL, extra_columns)
}
\arguments{
\item{input_filename}{The input (fasta) file, i.e. --infname argument}

\item{output_filename}{Desired output (csv) filename, i.e. --outfname 
argument}

\item{partis_path}{The full path to the partis executable}

\item{num_procs}{The number of processors to use in parallel}

\item{cleanup}{Flag to delete all interim files created by partis}

\item{output_path}{Desired output directory, which will contain 
\code{output_filename}}
}
\value{
A data.table object containing the output of the partis partition 
  call
}
\description{
Perform clonal partitioning with partis
}
