% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotAliphaticIndexDistribution}
\alias{plotAliphaticIndexDistribution}
\title{Plot the aliphatic index distribution of one or more datasets}
\usage{
plotAliphaticIndexDistribution(dat_list, do_exact = FALSE, names = NULL)
}
\arguments{
\item{dat_list}{A list of \code{data.table} objects corresponding to 
repertoire annotations}

\item{do_exact}{Display exact distribution plots rather than histograms}

\item{names}{Strings to be displayed by the legend corresponding to the 
elements of \code{dat_list}}
}
\description{
Plot the aliphatic index distribution of one or more datasets
}
