% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotCDR3LengthDistribution}
\alias{plotCDR3LengthDistribution}
\title{Plot the CDR3 length distribution of one or more datasets}
\usage{
plotCDR3LengthDistribution(dat_list, plot_types = c("freqpoly", "ecdf"),
  ...)
}
\arguments{
\item{dat_list}{A \code{list} of \code{data.table} objects corresponding to 
repertoire annotations}

\item{plot_types}{Vector of plot types to return. Can include "freqpoly",
"ecdf", or both}
}
\description{
Plot the CDR3 length distribution of one or more datasets
}
