% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotChargeDistribution}
\alias{plotChargeDistribution}
\title{Plot the charge distribution of one or more datasets}
\usage{
plotChargeDistribution(dat_list, plot_types = c("freqpoly", "ecdf"), ...)
}
\arguments{
\item{dat_list}{A \code{list} of \code{data.table} objects corresponding to 
repertoire annotations}

\item{plot_types}{Vector of plot types to return. Can include "freqpoly",
"ecdf", or both}
}
\description{
Plot the charge distribution of one or more datasets
}
