% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Plot a summary distribution of one or more datasets}
\usage{
plotDistribution(dat_list, summary_function, plot_type, x_label = "Value",
  names, show_legend = TRUE, ...)
}
\arguments{
\item{dat_list}{A \code{list} of \code{data.table} objects corresponding to 
repertoire annotations}

\item{summary_function}{A function that is applied to each dataset in 
\code{dat_list} and whose output values are plotted}

\item{x_label}{The text label for the x-axis}

\item{names}{Strings to be displayed by the legend corresponding to the 
elements of \code{dat_list}}

\item{do_exact}{Display exact distribution plots rather than histograms}
}
\description{
Plot a summary distribution of one or more datasets
}
