% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotUnivariateDistributions}
\alias{plotUnivariateDistributions}
\title{Generate a gridded plot of each univariate distribution corresponding to
  one or more annotations datasets}
\usage{
plotUnivariateDistributions(dat_list, tall_plot = FALSE, plot_type,
  names = NULL, plot_function_strings = NULL)
}
\arguments{
\item{dat_list}{A \code{list} of \code{data.table} objects corresponding to 
repertoire annotations}

\item{names}{Strings to be displayed by the legend corresponding to the 
elements of \code{dat_list}}
}
\description{
Generate a gridded plot of each univariate distribution corresponding to
  one or more annotations datasets
}
