% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotVGene3PrimeDeletionLengthDistribution}
\alias{plotVGene3PrimeDeletionLengthDistribution}
\title{Plot the V gene 3' deletion length distribution of one or more datasets}
\usage{
plotVGene3PrimeDeletionLengthDistribution(dat_list,
  plot_types = c("freqpoly", "ecdf"), ...)
}
\arguments{
\item{dat_list}{A \code{list} of \code{data.table} objects corresponding to 
repertoire annotations}

\item{plot_types}{Vector of plot types to return. Can include "freqpoly",
"ecdf", or both}
}
\description{
Plot the V gene 3' deletion length distribution of one or more datasets
}
