% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotVGene5PrimeDeletionLengthDistribution}
\alias{plotVGene5PrimeDeletionLengthDistribution}
\title{Plot the V gene 5' deletion length distribution of one or more datasets}
\usage{
plotVGene5PrimeDeletionLengthDistribution(dat_list, do_exact = FALSE,
  names = NULL)
}
\arguments{
\item{dat_list}{A list of \code{data.table} objects corresponding to 
repertoire annotations}

\item{do_exact}{Display exact distribution plots rather than histograms}

\item{names}{Strings to be displayed by the legend corresponding to the 
elements of \code{dat_list}}
}
\description{
Plot the V gene 5' deletion length distribution of one or more datasets
}
