% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotVJInsertionLengthDistribution}
\alias{plotVJInsertionLengthDistribution}
\title{This calls the function for VD insertions since the AIRR field
  for VD and VJ insertion lengths is the same}
\usage{
plotVJInsertionLengthDistribution(dat_list, plot_types = c("freqpoly",
  "ecdf"), ...)
}
\arguments{
\item{dat_list}{A \code{list} of \code{data.table} objects corresponding to 
repertoire annotations}

\item{plot_types}{Vector of plot types to return. Can include "freqpoly",
"ecdf", or both}
}
\description{
This calls the function for VD insertions since the AIRR field
  for VD and VJ insertion lengths is the same
}
