% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartisFunctions.R
\name{preventOutputOverwrite}
\alias{preventOutputOverwrite}
\title{Ensure sumrep does not accidentally delete any previous _output folders 
created by partis}
\usage{
preventOutputOverwrite(output_path, cleanup)
}
\arguments{
\item{output_path}{Desired partis output directory, which is \code{_output}
by default}

\item{cleanup}{Flag to delete all interim files created by partis}
}
\description{
Ensure sumrep does not accidentally delete any previous _output folders 
created by partis
}
