% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{removeBadAminoAcidSequences}
\alias{removeBadAminoAcidSequences}
\title{Remove strings in a list of amino acid sequences which contain at least one
  unrecognized amino acid or stop codon}
\usage{
removeBadAminoAcidSequences(aa_sequences)
}
\arguments{
\item{aa_sequences}{Vector or list of amino acid sequences}
}
\value{
Filtered list with "bad" sequences removed
}
\description{
Remove strings in a list of amino acid sequences which contain at least one
  unrecognized amino acid or stop codon
}
