% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{removeEmptyStrings}
\alias{removeEmptyStrings}
\title{Remove empty strings from a list or vector}
\usage{
removeEmptyStrings(l)
}
\arguments{
\item{l}{List or vector of strings}
}
\value{
The given list or vector with all empty strings removed
}
\description{
Remove empty strings from a list or vector
}
