% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare.R
\name{resample_data}
\alias{resample_data}
\title{Resample a dataset with replacement}
\usage{
resample_data(dat)
}
\arguments{
\item{dat}{data.table or data.frame object to be resampled}
}
\value{
A new, resampled data.table or data.frame object, depending on 
  the type of \code{dat}
}
\description{
\code{resample_data} returns a data.table (or data.frame) object with
  resampled rows from the input dataset. This yields a new dataset with the
  same number of rows and similar characteristics.
}
