% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartisFunctions.R
\name{simulateDataset}
\alias{simulateDataset}
\title{Simulate a dataset based on parameters from partis annotations}
\usage{
simulateDataset(parameter_dir, partis_path = Sys.getenv("PARTIS_PATH"),
  output_file = "simu.csv", num_events = NULL, num_leaves = NULL,
  cleanup = TRUE, do_full_annotation = TRUE)
}
\arguments{
\item{parameter_dir}{The parent output folder for partis (which is '_output'
by default. The function cd's into the params directory}

\item{num_events}{The desired number of VDJ rearragement events for the
simulation}
}
\description{
Simulate a dataset based on parameters from partis annotations
}
