% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{standardizeList}
\alias{standardizeList}
\title{Convert lists of factors or vectors of characters into a vector of strings}
\usage{
standardizeList(l)
}
\arguments{
\item{List}{or vector of either strings or char vectors of DNA sequences}
}
\value{
Vector of strings of DNA sequences
}
\description{
Convert lists of factors or vectors of characters into a vector of strings
}
