% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{subsampleVector}
\alias{subsampleVector}
\title{Subsample a vector}
\usage{
subsampleVector(v, sample_count)
}
\arguments{
\item{v}{Vector to subsample}

\item{sample_count}{Number of samples to retain in the subsampled data.
Samples refers to elements in a vector or rows in a data.table/data.frame}
}
\value{
A subsampled vector
}
\description{
Subsample a vector
}
