% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{subsetToQuantileRange}
\alias{subsetToQuantileRange}
\title{Subset a distribution to a given quantile range (mostly for plotting
  aesthetics)}
\usage{
subsetToQuantileRange(distribution, lower = 0.025, upper = 0.975)
}
\arguments{
\item{distribution}{The distribution of values to be subsetted}

\item{lower}{The lower quantile cutoff for values in \code{distribution}}

\item{upper}{The upper quantile cutoff for values in \code{distribution}}
}
\value{
A vector of values of \code{distribution} within the given quantiles
}
\description{
Subset a distribution to a given quantile range (mostly for plotting
  aesthetics)
}
