% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartisFunctions.R
\name{callPartis}
\alias{callPartis}
\title{Call partis}
\usage{
callPartis(action, input_filename, output_filename, output_path, partis_path,
  num_procs, germline_dir)
}
\arguments{
\item{action}{The desired partis command}

\item{input_filename}{The input (fasta) file, i.e. --infname argument}

\item{output_filename}{Desired output (csv) filename, i.e. --outfname 
argument}

\item{output_path}{Desired output directory, which will contain 
\code{output_filename}}

\item{partis_path}{The full path to the partis executable}

\item{num_procs}{The number of processors to use in parallel}

\item{cleanup}{Flag to delete all interim files created by partis}
}
\value{
A data.table object containing the output of the partis call
}
\description{
\code{callPartis} calls partis from within the sumrep package.
This is done through the bash script run_partis.sh.
It is assumed that a "SHELL" environmental variable is set to run the above 
script.
}
