% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareAtchleyFactorDistributions}
\alias{compareAtchleyFactorDistributions}
\title{Compare the distributions of Atchley factors for two datasets}
\usage{
compareAtchleyFactorDistributions(dat_a, dat_b, column = "junction_aa")
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire 
annotations}
}
\value{
Vector of divergences of each Atchley factor distribution
}
\description{
Compare the distributions of Atchley factors for two datasets
}
