% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareAtchleyFactorDistributions}
\alias{compareAtchleyFactorDistributions}
\title{Compare the distributions of mean Atchley factors for two datasets}
\usage{
compareAtchleyFactorDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\value{
Estimated mean absolute difference of the mean of each Atchley factor
  of \code{dat_a} and \code{dat_b}
}
\description{
Compare the distributions of mean Atchley factors for two datasets
}
