% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareCategoricalDistributions}
\alias{compareCategoricalDistributions}
\title{Get the l1 divergence of two categorical distributions, manifest as 
  \code{table} objects}
\usage{
compareCategoricalDistributions(d1, d2)
}
\arguments{
\item{d1, d2}{Table containing empirical frequencies}
}
\description{
Get the l1 divergence of two categorical distributions, manifest as 
  \code{table} objects
}
