% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareDGeneDistributions}
\alias{compareDGeneDistributions}
\title{Compare D gene distribution between two repertoires}
\usage{
compareDGeneDistributions(dat_a, dat_b)
}
\arguments{
\item{dat_a}{First annotated dataset}

\item{dat_b}{Second annotated dataset}
}
\value{
Mean absolute difference of D gene counts between the two
  repertoires
}
\description{
Compare D gene distribution between two repertoires
}
