% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareGermlineGeneDistributions}
\alias{compareGermlineGeneDistributions}
\title{Compare germline V, D, or J gene usage for two repertoires}
\usage{
compareGermlineGeneDistributions(dat_a, dat_b, gene_type, collapse_alleles,
  standardize)
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}

\item{gene_type}{String for gene type, taken as a column name of the 
annotated datasets. Must be "v_call", "d_call", or "j_call"}

\item{collapse_alleles}{Should allelic variants be ignored?}
}
\value{
l1 divergence of gene counts between the two repertoires
}
\description{
Compare germline V, D, or J gene usage for two repertoires
}
