% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareInFramePercentages}
\alias{compareInFramePercentages}
\title{Compare the percentage of in-frame sequences of two datasets}
\usage{
compareInFramePercentages(dat_a, dat_b, column = "vj_in_frame")
}
\arguments{
\item{dat_a, dat_b}{A \code{data.table} corresponding to repertoire annotations}
}
\value{
Absolute differences of in-frame percentages
}
\description{
Compare the percentage of in-frame sequences of two datasets
}
