% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareNNDistanceDistributions}
\alias{compareNNDistanceDistributions}
\title{Compare kth nearest neighbor distance distributions of two lists of
  sequences
\code{compareNNDistanceDistribution} computes the JS divergence of
  the kth nearest neighbor distance distribution of two lists of
  DNA sequences}
\usage{
compareNNDistanceDistributions(dat_a, dat_b, column = "mature_seq", k = 1,
  subsample_count = 100, trial_count = 10)
}
\arguments{
\item{dat_a}{First dataset, containing mature sequences}

\item{dat_b}{Second dataset, containing mature sequences}

\item{k}{The separation depth for the nearest neighbor distances.
k = 1 corresponds to the nearest neighbor, k = 2 corresponds to
the second-nearest neighbor, etc.}

\item{subsample_count}{Number of samples to extract on each trial}
}
\value{
Estimated JS divergence of the distributions inferred from list_a
  and list_b
}
\description{
Compare kth nearest neighbor distance distributions of two lists of
  sequences
\code{compareNNDistanceDistribution} computes the JS divergence of
  the kth nearest neighbor distance distribution of two lists of
  DNA sequences
}
