% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{comparePairwiseDistanceDistributions}
\alias{comparePairwiseDistanceDistributions}
\title{Compare pairwise distance distributions of two lists of sequences}
\usage{
comparePairwiseDistanceDistributions(list_a, list_b)
}
\arguments{
\item{list_a}{First list of sequences}

\item{list_b}{Second list of sequences}
}
\value{
JS divergence of the distributions inferred from list_a
  and list_b
}
\description{
\code{comparePairwiseDistanceDistributions} computes the JS
  divergence of the pairwise distance distribution of two lists of
  DNA sequences
}
