% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare.R
\name{compareRepertoires}
\alias{compareRepertoires}
\title{Run through a full repertoire comparison}
\usage{
compareRepertoires(repertoire_1, repertoire_2, rep_1_bootstrap = NULL,
  receptor_type, chain_type, do_full_comparison = FALSE)
}
\arguments{
\item{repertoire_1, repertoire_2}{List including a \code{data.table} named
\code{annotations} and optionally a list called \code{mutation_rates}}

\item{rep_1_bootstrap}{Annotated repertoire based on bootstrapping the DNA
sequences from the first repertoire}

\item{receptor_type}{A string denoting the type of immune receptor to which
\code{repertoire_1} and \code{repertoire_2} correspond.
Either "BCR" or "TCR".}

\item{chain_type}{The locus from which the sequences were sampled. Either
"heavy", "light", "beta", or "alpha".}

\item{do_full_comparison}{If TRUE, performs all available comparisons for
the given receptor_type/chain_type combo, including comparisons for the 
slower summaries}
}
\description{
\code{compareRepertoires} iterates through the various comparison functions
  in \code{sumrep}, printing the distance or divergence value of each one.
  Both repertoires are assumed to be annotated.
  Partition-based comparisons are available if the repertoires are 
  partitioned.
  Comparisons based on per-gene and per-gene-per-position mutation rates are 
  also available given mutation rate information.
}
