% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare.R
\name{compareRepertoires}
\alias{compareRepertoires}
\title{Run through a full repertoire comparison}
\usage{
compareRepertoires(repertoire_1, repertoire_2, rep_1_bootstrap = NULL,
  locus, do_full_comparison = FALSE)
}
\arguments{
\item{repertoire_1, repertoire_2}{List including a \code{data.table} named
\code{annotations} and optionally a list called \code{mutation_rates}}

\item{rep_1_bootstrap}{Annotated repertoire based on bootstrapping the DNA
sequences from the first repertoire}

\item{locus}{String denoting the locus for the receptors given in 
\code{input_filename}. 
Either "tra", "trb", "trd", or "trg" for TCRs,
or "igl", "igk", or "igh" for BCRs}

\item{do_full_comparison}{If TRUE, performs all available comparisons for
the given locus, including comparisons for the slower summaries}
}
\description{
\code{compareRepertoires} iterates through the various comparison functions
  in \code{sumrep}, printing the distance or divergence value of each one.
  Both repertoires are assumed to be annotated.
  Partition-based comparisons are available if the repertoires are 
  partitioned.
  Comparisons based on per-gene and per-gene-per-position mutation rates are 
  also available given mutation rate information.
}
