% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareVDInsertionMatrices}
\alias{compareVDInsertionMatrices}
\title{Compare the transition matrices for VD insertions for two datasets}
\usage{
compareVDInsertionMatrices(dat_a, dat_b)
}
\arguments{
\item{dat_a}{First data.table with insertion annotations}

\item{dat_b}{First data.table with insertion annotations}
}
\value{
The mean absolute difference of matrix entries, taken elementwise
}
\description{
Compare the transition matrices for VD insertions for two datasets
}
