% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareVDJDistributions}
\alias{compareVDJDistributions}
\title{Compare joint V, D, and J gene usage between two annotated repertoires}
\usage{
compareVDJDistributions(dat_a, dat_b, collapseAlleles = TRUE)
}
\arguments{
\item{dat_a}{First annotated dataset}

\item{dat_b}{Second annotated dataset}
}
\value{
Mean absolute difference between joint gene usage, over all
  observed genes between the two repertoires
}
\description{
Compare joint V, D, and J gene usage between two annotated repertoires
}
