% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{compareVGeneDistributions}
\alias{compareVGeneDistributions}
\title{Compare V gene distribution between two repertoires}
\usage{
compareVGeneDistributions(dat_a, dat_b, collapse_alleles = TRUE)
}
\arguments{
\item{dat_a}{A \code{data.table} corresponding to repertoire annotations}

\item{dat_b}{A \code{data.table} corresponding to repertoire annotations}

\item{collapse_alleles}{Should allelic variants be ignored?}
}
\value{
l1 divergence of V gene counts between the two repertoires
}
\description{
Compare V gene distribution between two repertoires
}
