% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare.R
\name{doComparison}
\alias{doComparison}
\title{Apply \code{function_string} to inputs \code{input_1} and \code{input_2},
  and print the results}
\usage{
doComparison(function_string, input_1, input_2)
}
\arguments{
\item{function_string}{The name of the function to apply. Must be a 
comparison that takes two inputs}

\item{input_1}{First input to \code{function_string}}

\item{input_2}{Second input to \code{function_string}}
}
\description{
Apply \code{function_string} to inputs \code{input_1} and \code{input_2},
  and print the results
}
