% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getAliphaticIndexDistribution}
\alias{getAliphaticIndexDistribution}
\title{Get the distribution of aliphatic indices of a list of sequences}
\usage{
getAliphaticIndexDistribution(dat, column = "junction_aa")
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}
}
\value{
Vector of aliphatic indices
}
\description{
Get the distribution of aliphatic indices of a list of sequences
}
