% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare.R
\name{getAndPrintComparison}
\alias{getAndPrintComparison}
\title{Print the result and elapsed time of a comparison}
\usage{
getAndPrintComparison(f, input_1, input_2, string_header, color)
}
\arguments{
\item{f}{Comparison function}

\item{input_1}{First input to \code{f}}

\item{input_2}{Second input to \code{f}}

\item{string_header}{Text to precede the comparison value and elapsed time}

\item{color}{Color for text. Green for usual comparison, yellow for 
bootstrap}
}
\description{
Print the result and elapsed time of a comparison
}
