% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approximation.R
\name{getApproximateNearestNeighborDistribution}
\alias{getApproximateNearestNeighborDistribution}
\title{Runs a special approximation routine since nearest neighbor distances are
  not a regular distribution for subsampling. (Since the nearest neighbor
  distance of a sample is necessarily larger than that of the population)}
\usage{
getApproximateNearestNeighborDistribution(dat, column, tol = 0.001,
  batch_size = 50)
}
\arguments{
\item{dat}{}

\item{column}{The column over which the nearest neighbor distance
should be computed. Must correspond to sequence strings, e.g. DNA or
amino acid sequences.}

\item{tol}{The allowed tolerance for determining convergence.}

\item{batch_size}{The number of sequences sampled during each iteration.}

\item{dat}{The dataset for which the distribution is desired}
}
\value{
Approximate nearest neighbor distance distribution of sequences
  given by \code{dat[[column]]}
}
\description{
Runs a special approximation routine since nearest neighbor distances are
  not a regular distribution for subsampling. (Since the nearest neighbor
  distance of a sample is necessarily larger than that of the population)
}
