% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getDistancesFromNaiveToMature}
\alias{getDistancesFromNaiveToMature}
\title{Get distribution of Levenshtein distances from the inferred naive sequences
to the observed mature ones, sequence by sequence}
\usage{
getDistancesFromNaiveToMature(dat)
}
\arguments{
\item{dat}{Annotated dataset}
}
\value{
Vector of Levenshtein distances from naive to mature
}
\description{
Get distribution of Levenshtein distances from the inferred naive sequences
to the observed mature ones, sequence by sequence
}
