% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getGCContentDistribution}
\alias{getGCContentDistribution}
\title{Get the GC content distribution of a list of DNA sequences}
\usage{
getGCContentDistribution(dat, column = "sequence_alignment",
  approximate = FALSE, ...)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}

\item{approximate}{If TRUE, approximate distribution by subsampling.}
}
\value{
A vector of GC content values
}
\description{
\code{getGCContentDistribution} returns a list of the GC content of each DNA
  sequence in \code{raw_sequences}, given by the \code{ape} library
}
