% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getGRAVYDistribution}
\alias{getGRAVYDistribution}
\title{Get the distribution of GRAVY values from a list or vector of sequences}
\usage{
getGRAVYDistribution(sequence_list)
}
\arguments{
\item{List}{or vector of DNA sequence strings}
}
\value{
Vector of GRAVY values for \code{sequence_list}
}
\description{
Get the distribution of GRAVY values from a list or vector of sequences
}
