% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgBlastFunctions.R
\name{getIgBlastAnnotations}
\alias{getIgBlastAnnotations}
\title{Get IgBLAST annotations from an input fasta file}
\usage{
getIgBlastAnnotations(input_filename,
  output_filename = "igblast_out.tsv", organism = "human",
  domain_system = "imgt", locus, nproc = 4,
  igblast_dir = Sys.getenv("IGBLAST_DIR"),
  changeo_dir = Sys.getenv("CHANGEO_DIR"), cleanup = TRUE)
}
\arguments{
\item{input_filename}{The path to the input fasta file}

\item{output_filename}{The desired output filename (should be a tsv file)}

\item{organism}{The organism/species to which the sequences belong}

\item{domain_system}{(input to IgBLAST -- alter with caution)}

\item{locus}{String denoting the locus for the receptors given in 
\code{input_filename}. 
Either "tra", "trb", "trd", or "trg" for TCRs,
or "igl", "igk", or "igh" for BCRs}

\item{nproc}{The number of processors for use within the igblast and
shazam::distToNearest function calls. Defaults to 4.
Setting nproc=alakazam::cpuCount() attempts to automatically detect the
number of available cores -- see the alakazam documtation for details.}

\item{igblast_dir}{Path to parent directory of the igblast bin folder
(e.g. "path/to/parent", NOT "path/to/parent/bin").}

\item{changeo_dir}{The path of the changeo bin folder (e.g.
"path/to/bin").}

\item{cleanup}{If TRUE, remove all interim files created}
}
\value{
A \code{list} of a \code{data.table} containing the annotations.
}
\description{
The routine changes directories to the provided igblast directory,
  runs some Change-O commands, reads in the resultant tsv files,
  changes back to the initial directory, and returns the annotations
  list object.
}
