% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgBlastFunctions.R
\name{getIgBlastAnnotations}
\alias{getIgBlastAnnotations}
\title{Get IgBLAST annotations from an input fasta file}
\usage{
getIgBlastAnnotations(input_filename, output_filename = "igblast_out.tsv",
  organism = "human", domain_system = "imgt", receptor_type,
  num_threads = 8, igblast_dir, changeo_dir, cleanup = TRUE)
}
\arguments{
\item{input_filename}{The path to the input fasta file}

\item{output_filename}{The desired output filename (should be a tsv file)}

\item{organism}{The organism/species to which the sequences belong}

\item{domain_system}{(input to IgBLAST -- alter with caution)}

\item{receptor_type}{String denoting the type of receptors given in 
\code{input_filename}. Either "BCR" or "TCR".}

\item{igblast_dir}{Path to parent directory of the igblast bin folder
(e.g. "path/to/parent", NOT "path/to/parent/bin").}

\item{changeo_dir}{The path of the changeo bin folder (e.g.
"path/to/bin").}

\item{cleanup}{If TRUE, remove all interim files created}
}
\value{
A \code{list} of a \code{data.table} containing the annotations.
}
\description{
The routine changes directories to the provided igblast directory,
  runs some Change-O commands, reads in the resultant tsv files,
  changes back to the initial directory, and returns the annotations
  list object.
}
