% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getJointGeneTable}
\alias{getJointGeneTable}
\title{Get table of combined V, D, and J usage frequencies}
\usage{
getJointGeneTable(dat, collapseAlleles)
}
\arguments{
\item{dat}{Annotated dataset}
}
\value{
A table of joint gene IDs and usage counts
}
\description{
Get table of combined V, D, and J usage frequencies
}
