% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getJointGeneTable}
\alias{getJointGeneTable}
\title{Get table of combined V, D, and J usage frequencies}
\usage{
getJointGeneTable(dat, collapse_alleles)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{collapse_alleles}{Should allelic variants be ignored?}
}
\value{
A table of joint gene IDs and usage counts
}
\description{
Get table of combined V, D, and J usage frequencies
}
