% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getMeanAtchleyFactorDistribution}
\alias{getMeanAtchleyFactorDistribution}
\title{Get the distribution of mean Atchley factors for a list of DNA seuqnces}
\usage{
getMeanAtchleyFactorDistribution(sequence_list)
}
\arguments{
\item{sequence_list}{List or vector of DNA sequences}
}
\value{
Vector of mean Atchley factors of \code{sequence_list}
}
\description{
Get the distribution of mean Atchley factors for a list of DNA seuqnces
}
