% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getNearestNeighborDistribution}
\alias{getNearestNeighborDistribution}
\title{Get exact or approximate nearest neighbor distance distribution}
\usage{
getNearestNeighborDistribution(dat, column = "sequence", k = 1,
  approximate = TRUE, ...)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}

\item{k}{The separation depth for the nearest neighbor distances.
k = 1 corresponds to the nearest neighbor, k = 2 corresponds to
the second-nearest neighbor, etc.}

\item{approximate}{If TRUE, approximate distribution by subsampling.
Note: this inhibits interpretability since any subsample will have a 
different definition of a "nearest neighbor"}
}
\value{
vector of integer-value distances
}
\description{
Get exact or approximate nearest neighbor distance distribution
}
