% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getPairwiseDistanceDistribution}
\alias{getPairwiseDistanceDistribution}
\title{Get an exact or approximate distribution of pairwise distances}
\usage{
getPairwiseDistanceDistribution(dat, column = "sequence",
  approximate = TRUE, ...)
}
\arguments{
\item{dat}{A \code{data.table} corresponding to repertoire annotations}

\item{column}{the column name of \code{dat} containing the strings on which
the distribution should be computed}

\item{approximate}{if TRUE, approximate the distribution by subsampling
and averaging}
}
\value{
vector of integer-valued distances
}
\description{
Get an exact or approximate distribution of pairwise distances
}
