% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{getSubstitutionModel}
\alias{getSubstitutionModel}
\title{Get the inferred substitution model of a repertoire}
\usage{
getSubstitutionModel(dat)
}
\arguments{
\item{dat}{Annotated dataset}
}
\value{
The default shazam substitution model
}
\description{
Get the inferred substitution model of a repertoire
}
