% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotChargeDistribution}
\alias{plotChargeDistribution}
\title{Plot the charge distribution of one or more datasets}
\usage{
plotChargeDistribution(dat_list, plot_type, names = NULL,
  show_legend = TRUE, ...)
}
\arguments{
\item{dat_list}{A \code{list} of \code{data.table} objects corresponding to 
repertoire annotations}

\item{names}{Strings to be displayed by the legend corresponding to the 
elements of \code{dat_list}}
}
\description{
Plot the charge distribution of one or more datasets
}
