% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Plot a summary distribution of one or more datasets}
\usage{
plotDistribution(dat_list, summary_function, plot_types,
  x_label = "Value", color = NULL, lty = NULL, bins = 15, ...)
}
\arguments{
\item{dat_list}{A \code{list} of \code{data.table} objects corresponding to 
repertoire annotations}

\item{summary_function}{A function that is applied to each dataset in 
\code{dat_list} and whose output values are plotted}

\item{plot_types}{Vector of plot types to return. Can include "freqpoly",
"ecdf", or both}

\item{x_label}{The text label for the x-axis}

\item{color}{Vector of names by which to group \code{dat_list} by color}

\item{lty}{Vector of names by which to group \code{dat_list} by linetype}

\item{bins}{Number of bins to group the frequency polygon}
}
\value{
List of plots of the given summary for each given dataset
}
\description{
Plot a summary distribution of one or more datasets
}
