% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotUnivariateDistributions}
\alias{plotUnivariateDistributions}
\title{Generate a gridded plot of each univariate distribution corresponding to
  one or more annotations datasets}
\usage{
plotUnivariateDistributions(dat_list, plot_types = c("freqpoly", "ecdf"),
  locus, plot_function_strings = NULL, functions_to_omit = NULL,
  do_all_plots = FALSE, color = factor(1:length(dat_list)),
  lty = factor(rep(1, length(dat_list))), bins = 10,
  use_viridis = FALSE)
}
\arguments{
\item{dat_list}{List of datasets}

\item{plot_types}{Vector of plot types for each distribution.
Currently, only "freqpoly" and "ecdf" are supported.}

\item{locus}{The locus to which the datasets in \code{dat_list} correspond}
}
\value{
A list of ggplot objects containing the specified plots
}
\description{
Generate a gridded plot of each univariate distribution corresponding to
  one or more annotations datasets
}
