% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{plotVGene3PrimeDeletionLengthDistribution}
\alias{plotVGene3PrimeDeletionLengthDistribution}
\title{Plot the V gene 3' deletion length distribution of one or more datasets}
\usage{
plotVGene3PrimeDeletionLengthDistribution(dat_list, plot_type, names = NULL,
  show_legend = TRUE)
}
\arguments{
\item{dat_list}{A \code{list} of \code{data.table} objects corresponding to 
repertoire annotations}

\item{names}{Strings to be displayed by the legend corresponding to the 
elements of \code{dat_list}}
}
\description{
Plot the V gene 3' deletion length distribution of one or more datasets
}
